// src/maocars-client/schemas/carBodyStyle.ts
/* eslint-disable */
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */

export type CarBodyStyle = (typeof CarBodyStyle)[keyof typeof CarBodyStyle];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarBodyStyle = {
  coupe: 'coupe',
  convertible: 'convertible',
  hatchback: 'hatchback',
  sedan: 'sedan',
  suv: 'suv',
  truck: 'truck',
  van: 'van',
  wagon: 'wagon',
} as const;