// src/maocars-client/schemas/carDriveTrain.ts
/* eslint-disable */
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */

export type CarDriveTrain = (typeof CarDriveTrain)[keyof typeof CarDriveTrain];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarDriveTrain = {
  front: 'front',
  rear: 'rear',
  'all-drive': 'all-drive',
} as const;