// src/maocars-client/schemas/carSellerType.ts
/* eslint-disable */
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */

export type CarSellerType = (typeof CarSellerType)[keyof typeof CarSellerType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarSellerType = {
  private: 'private',
  dealer: 'dealer',
} as const;