// src/maocars-client/schemas/carTransmission.ts
/* eslint-disable */
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * mao-cars-api
 */

export type CarTransmission = (typeof CarTransmission)[keyof typeof CarTransmission];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CarTransmission = {
  automatic: 'automatic',
  manual: 'manual',
} as const;